pyproject.toml
[project]
name = "pubmed-fetcher"
version = "0.1.0"
description = "A Python tool to fetch research papers from PubMed."
authors = [
    {name = "Nunna Swathi",email = "swathinunna11@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
